Refactor anti-censorship into seperate module
