--- android/app/src/main/kotlin/net/mullvad/mullvadvpn/compose/state/Udp2TcpSettingsUiState.kt
@@ -1,6 +1,0 @@
-package net.mullvad.mullvadvpn.compose.state
-
-import net.mullvad.mullvadvpn.lib.model.Constraint
-import net.mullvad.mullvadvpn.lib.model.Port
-
-data class Udp2TcpSettingsUiState(val port: Constraint<Port> = Constraint.Any)
