+++ android/lib/feature/anticensorship/impl/src/main/java/net/mullvad/mullvadvpn/anticensorship/impl/AntiCensorshipUiStatePreviewParameterProvider.kt
@@ -1,0 +1,23 @@
+package net.mullvad.mullvadvpn.anticensorship.impl
+
+import androidx.compose.ui.tooling.preview.PreviewParameterProvider
+import net.mullvad.mullvadvpn.lib.common.Lc
+import net.mullvad.mullvadvpn.lib.common.toLc
+import net.mullvad.mullvadvpn.lib.model.Constraint
+import net.mullvad.mullvadvpn.lib.model.ObfuscationMode
+
+class AntiCensorshipUiStatePreviewParameterProvider :
+    PreviewParameterProvider<Lc<Boolean, AntiCensorshipSettingsUiState>> {
+    override val values =
+        sequenceOf(
+            AntiCensorshipSettingsUiState.from(
+                    isModal = false,
+                    selectedWireguardPort = Constraint.Any,
+                    obfuscationMode = ObfuscationMode.Udp2Tcp,
+                    selectedUdp2TcpObfuscationPort = Constraint.Any,
+                    selectedShadowsocksObfuscationPort = Constraint.Any,
+                )
+                .toLc(),
+            Lc.Loading(true),
+        )
+}
