+++ android/lib/feature/anticensorship/impl/src/main/java/net/mullvad/mullvadvpn/anticensorship/impl/customport/CustomPortDialog.kt
@@ -1,0 +1,133 @@
+package net.mullvad.mullvadvpn.anticensorship.impl.customport
+
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.testTag
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.buildAnnotatedString
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import androidx.lifecycle.compose.dropUnlessResumed
+import com.ramcosta.composedestinations.annotation.Destination
+import com.ramcosta.composedestinations.annotation.ExternalModuleGraph
+import com.ramcosta.composedestinations.result.ResultBackNavigator
+import com.ramcosta.composedestinations.spec.DestinationStyle
+import net.mullvad.mullvadvpn.common.compose.CollectSideEffectWithLifecycle
+import net.mullvad.mullvadvpn.lib.common.util.asString
+import net.mullvad.mullvadvpn.lib.model.Port
+import net.mullvad.mullvadvpn.lib.model.PortRange
+import net.mullvad.mullvadvpn.lib.model.PortType
+import net.mullvad.mullvadvpn.lib.ui.component.annotatedStringResource
+import net.mullvad.mullvadvpn.lib.ui.component.dialog.InputDialog
+import net.mullvad.mullvadvpn.lib.ui.resource.R
+import net.mullvad.mullvadvpn.lib.ui.tag.CUSTOM_PORT_DIALOG_INPUT_TEST_TAG
+import net.mullvad.mullvadvpn.lib.ui.theme.AppTheme
+import org.koin.androidx.compose.koinViewModel
+
+@Preview
+@Composable
+private fun PreviewWireguardCustomPortDialog() {
+    AppTheme {
+        CustomPortDialog(
+            title = "Custom port",
+            portInput = "",
+            isValidInput = false,
+            allowedPortRanges = listOf(PortRange(10..10), PortRange(40..50)),
+            recommendedPortRanges = listOf(PortRange(10..10)),
+            showResetToDefault = false,
+            onInputChanged = {},
+            onSavePort = {},
+            onResetPort = {},
+            onDismiss = {},
+        )
+    }
+}
+
+@Destination<ExternalModuleGraph>(style = DestinationStyle.Dialog::class)
+@Composable
+fun CustomPort(navArg: CustomPortDialogNavArgs, backNavigator: ResultBackNavigator<Port?>) {
+    val viewModel = koinViewModel<CustomPortDialogViewModel>()
+
+    val uiState by viewModel.uiState.collectAsStateWithLifecycle()
+
+    CollectSideEffectWithLifecycle(viewModel.uiSideEffect) {
+        when (it) {
+            is CustomPortDialogSideEffect.Success -> backNavigator.navigateBack(it.port)
+        }
+    }
+
+    val title =
+        when (navArg.portType) {
+            PortType.Udp2Tcp -> stringResource(R.string.udp_over_tcp)
+            PortType.Shadowsocks -> stringResource(R.string.shadowsocks)
+            PortType.Wireguard -> stringResource(R.string.wireguard)
+            PortType.Lwo -> stringResource(R.string.lwo)
+        }
+
+    CustomPortDialog(
+        title = title,
+        portInput = uiState.portInput,
+        isValidInput = uiState.isValidInput,
+        showResetToDefault = uiState.showResetToDefault,
+        allowedPortRanges = uiState.allowedPortRanges,
+        recommendedPortRanges = uiState.recommendedPortRanges,
+        onInputChanged = viewModel::onInputChanged,
+        onSavePort = viewModel::onSaveClick,
+        onResetPort = viewModel::onResetClick,
+        onDismiss = dropUnlessResumed { backNavigator.navigateBack() },
+    )
+}
+
+@Composable
+fun CustomPortDialog(
+    title: String,
+    portInput: String,
+    isValidInput: Boolean,
+    allowedPortRanges: List<PortRange>,
+    recommendedPortRanges: List<PortRange>,
+    showResetToDefault: Boolean,
+    onInputChanged: (String) -> Unit,
+    onSavePort: (String) -> Unit,
+    onResetPort: () -> Unit,
+    onDismiss: () -> Unit,
+) {
+    InputDialog(
+        title = title,
+        message =
+            buildAnnotatedString {
+                appendLine(
+                    stringResource(
+                        id = R.string.custom_port_dialog_valid_ranges,
+                        allowedPortRanges.asString(),
+                    )
+                )
+                if (recommendedPortRanges.isNotEmpty()) {
+                    append(annotatedStringResource(R.string.custom_port_recommended_range_first))
+                    append(
+                        stringResource(
+                            id = R.string.custom_port_recommended_range_second,
+                            recommendedPortRanges.asString(),
+                        )
+                    )
+                }
+            },
+        confirmButtonEnabled = isValidInput,
+        confirmButtonText = stringResource(id = R.string.custom_port_dialog_submit),
+        onResetButtonText = stringResource(R.string.custom_port_dialog_remove),
+        onBack = onDismiss,
+        onConfirm = { onSavePort(portInput) },
+        onReset = if (showResetToDefault) onResetPort else null,
+        input = {
+            CustomPortTextField(
+                value = portInput,
+                onValueChanged = onInputChanged,
+                onSubmit = onSavePort,
+                isValidValue = isValidInput,
+                maxCharLength = 5,
+                modifier = Modifier.testTag(CUSTOM_PORT_DIALOG_INPUT_TEST_TAG).fillMaxWidth(),
+            )
+        },
+    )
+}
