+++ android/lib/feature/anticensorship/impl/src/main/java/net/mullvad/mullvadvpn/anticensorship/impl/customport/CustomPortDialogSideEffect.kt
@@ -1,0 +1,7 @@
+package net.mullvad.mullvadvpn.anticensorship.impl.customport
+
+import net.mullvad.mullvadvpn.lib.model.Port
+
+sealed interface CustomPortDialogSideEffect {
+    data class Success(val port: Port?) : CustomPortDialogSideEffect
+}
