+++ android/lib/feature/anticensorship/impl/src/main/java/net/mullvad/mullvadvpn/anticensorship/impl/customport/CustomPortDialogUiState.kt
@@ -1,0 +1,11 @@
+package net.mullvad.mullvadvpn.anticensorship.impl.customport
+
+import net.mullvad.mullvadvpn.lib.model.PortRange
+
+data class CustomPortDialogUiState(
+    val portInput: String,
+    val isValidInput: Boolean,
+    val allowedPortRanges: List<PortRange>,
+    val recommendedPortRanges: List<PortRange>,
+    val showResetToDefault: Boolean,
+)
