+++ android/lib/feature/anticensorship/impl/src/main/java/net/mullvad/mullvadvpn/anticensorship/impl/selectport/SelectPortNavArgs.kt
@@ -1,0 +1,7 @@
+package net.mullvad.mullvadvpn.anticensorship.impl.selectport
+
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+import net.mullvad.mullvadvpn.lib.model.PortType
+
+@Parcelize data class SelectPortNavArgs(val portType: PortType) : Parcelable
