+++ android/lib/feature/anticensorship/impl/src/main/java/net/mullvad/mullvadvpn/anticensorship/impl/selectport/SelectPortUiState.kt
@@ -1,0 +1,31 @@
+package net.mullvad.mullvadvpn.anticensorship.impl.selectport
+
+import com.ramcosta.composedestinations.spec.BaseRoute
+import com.ramcosta.composedestinations.spec.Direction
+import net.mullvad.mullvadvpn.lib.model.Constraint
+import net.mullvad.mullvadvpn.lib.model.Port
+import net.mullvad.mullvadvpn.lib.model.PortRange
+import net.mullvad.mullvadvpn.lib.model.PortType
+
+data class SelectPortUiState(
+    val portType: PortType,
+    val port: Constraint<Port> = Constraint.Any,
+    val customPort: Port? = null,
+    val customPortEnabled: Boolean,
+    val title: String,
+    val allowedPortRanges: List<PortRange> = emptyList(),
+    val recommendedPortRanges: List<PortRange> = emptyList(),
+    val presetPorts: List<Port> = emptyList(),
+    val infoDestination: BaseRoute? = null,
+) {
+    val isCustom = port is Constraint.Only && port.value !in presetPorts
+}
+
+data class PortTypeUiState(
+    val presetPorts: List<Port>,
+    val allowedPortRanges: List<PortRange>,
+    val recommendedPortRanges: List<PortRange>,
+    val customPortEnabled: Boolean,
+    val title: String,
+    val infoDestination: Direction? = null,
+)
