+++ android/settings.gradle.kts
@@ -38,6 +38,7 @@ include(
     ":lib:common-test",
     ":lib:grpc",
     ":lib:endpoint",
+    ":lib:feature:anticensorship:impl",
     ":lib:feature:appinfo:impl",
     ":lib:feature:appearance:impl",
     ":lib:feature:daita:impl",
